+++ __tests__/serializers/slateToHtml/index.spec.ts
@@ -1,6 +1,9 @@
 import { ChildNode, Element } from 'domhandler'
 import { slateToHtml, slateToDomConfig } from '../../../src'
 
+const postcss = require('postcss')
+const postcssJs = require('postcss-js')
+
 describe('slateToHtml expected behaviour', () => {
   it('encodes HTML entities', () => {
     const html = `<h1>What&apos;s Heading 1</h1>`
@@ -15,6 +18,68 @@ describe('slateToHtml expected behaviour', () => {
       },
     ]
     expect(slateToHtml(slate)).toEqual(html)
+  })
+
+  /**
+   * @see https://www.w3.org/International/questions/qa-escapes#use
+   */
+  it('encodes `breaking` HTML entities', () => {
+    const html = `<p>2 &gt; 1 but is &lt; 3 &amp; it can break HTML</p>`
+    const slate = [
+      {
+        children: [
+          {
+            text: '2 > 1 but is < 3 & it can break HTML',
+          },
+        ],
+        type: 'p',
+      },
+    ]
+    expect(slateToHtml(slate)).toEqual(html)
+  })
+
+  it('encodes `non breaking` HTML entities', () => {
+    const html = `<p>The company&#x2019;s priority is &apos;inside sales&apos; and changing the spelling of cafe to caf&#xe9;.</p>`
+    const slate = [
+      {
+        children: [
+          {
+            text: "The company’s priority is 'inside sales' and changing the spelling of cafe to café.",
+          },
+        ],
+        type: 'p',
+      },
+    ]
+    expect(slateToHtml(slate)).toEqual(html)
+  })
+
+  it('encodes `breaking` HTML entities only if option is active', () => {
+    const html = `<p>2 &gt; 1 but is &lt; 3 &amp; it can break HTML</p><p>The company’s priority is 'inside sales' and changing the spelling of cafe to café.</p>`
+    const slate = [
+      {
+        children: [
+          {
+            text: '2 > 1 but is < 3 & it can break HTML',
+          },
+        ],
+        type: 'p',
+      },
+      {
+        children: [
+          {
+            text: "The company’s priority is 'inside sales' and changing the spelling of cafe to café.",
+          },
+        ],
+        type: 'p',
+      },
+    ]
+    expect(
+      slateToHtml(slate, {
+        ...slateToDomConfig,
+        encodeEntities: false,
+        alwaysEncodeBreakingEntities: true,
+      }),
+    ).toEqual(html)
   })
 
   it('does not encode HTML entities with the appropriate option', () => {
@@ -68,7 +133,9 @@ describe('slateToHtml expected behaviour', () => {
     ]
     expect(slateToHtml(slate)).toEqual(html)
   })
+})
 
+describe('custom config', () => {
   it('respects the alwaysEncodeCodeEntities option if encodeEntities is false', () => {
     const html = '<p>Regular text & <pre><code>&lt;textarea&gt;</code></pre>.</p>'
     const slate = [
@@ -93,7 +160,28 @@ describe('slateToHtml expected behaviour', () => {
     )
   })
 
-  it('process a custom element tag map', () => {
+  it('processes an element map value', () => {
+    const html = '<h1>Heading 1</h1>'
+    const slate = [
+      {
+        type: 'heading-one',
+        children: [
+          {
+            text: 'Heading 1',
+          },
+        ],
+      },
+    ]
+    const config = {
+      ...slateToDomConfig,
+      elementMap: {
+        ['heading-one']: 'h1'
+      },
+    }
+    expect(slateToHtml(slate, config)).toEqual(html)
+  })
+
+  it('processes an element transform', () => {
     const html = '<p>Paragraph</p><img src="https://picsum.photos/id/237/200/300">'
     const slate = [
       {
@@ -116,6 +204,228 @@ describe('slateToHtml expected behaviour', () => {
         image: ({ node }: { node?: any }) => {
           return new Element('img', {
             src: node.url,
+          })
+        },
+      },
+    }
+    expect(slateToHtml(slate, config)).toEqual(html)
+  })
+
+  it('maps Slate attribute to inline style from element style map', () => {
+    const html = '<p style="font-size:96px;"><strong>Paragraph</strong></p>'
+    const slate = [
+      {
+        type: 'p',
+        fontSize: '96px',
+        children: [
+          {
+            bold: true,
+            text: 'Paragraph',
+          },
+        ],
+      }
+    ]
+    const config = {
+      ...slateToDomConfig,
+      elementStyleMap: {
+        fontSize: 'font-size',
+      }
+    }
+    expect(slateToHtml(slate, config)).toEqual(html)
+  })
+
+  it('processes a mark map value', () => {
+    const html = '<p><sub>Subscript text</sub></p>'
+    const slate = [
+      {
+        type: 'p',
+        children: [
+          {
+            text: 'Subscript text',
+            subScript: true,
+          },
+        ],
+      },
+    ]
+    const config = {
+      ...slateToDomConfig,
+      markMap: {
+        subScript: ['sub']
+      },
+    }
+    expect(slateToHtml(slate, config)).toEqual(html)
+  })
+
+  it('processes a mark transform', () => {
+    const html = '<p><strong style="font-size:96px;">Paragraph</strong></p>'
+    const slate = [
+      {
+        type: 'p',
+        children: [
+          {
+            bold: true,
+            fontSize: '96px',
+            text: 'Paragraph',
+          },
+        ],
+      }
+    ]
+    const config = {
+      ...slateToDomConfig,
+      markTransforms: {
+        ...slateToDomConfig.markTransforms,
+        strong: ({ node }: { node?: any }) => {
+          return new Element('strong', {
+            style: `font-size:${node.fontSize};`,
+          })
+        },
+      },
+    }
+    expect(slateToHtml(slate, config)).toEqual(html)
+  })
+})
+
+describe('style attribute css transforms with postcss', () => {
+  const transformStyleObjectToString = (style: {[key: string]: any}) => {
+    const postcssOptions = {
+      parser: postcssJs,
+      from: undefined, 
+    }
+    return postcss().process(style, postcssOptions).css.replace(/(\r\n|\n|\r)/gm, " ").replace(/\s\s+/g, ' ')
+  }
+
+  it('element transforms', () => {
+    const html = '<p style=\"font-size: 96px; --text-color: #DD3A0A; @media screen { z-index: 1; color: var(--text-color) }\"><strong>Paragraph</strong></p>'
+    const slate = [
+      {
+        type: 'p',
+        style: {
+          fontSize: '96px',
+          '--text-color': '#DD3A0A',
+          '@media screen': {
+            zIndex: '1',
+            color: 'var(--text-color)'
+          }
+        },
+        children: [
+          {
+            bold: true,
+            text: 'Paragraph',
+          },
+        ],
+      }
+    ]
+    const config = {
+      ...slateToDomConfig,
+      elementTransforms: {
+        ...slateToDomConfig.elementTransforms,
+        p: ({ node, children }: { node?: any, children?: ChildNode[] }) => {
+          return new Element('p', {
+            style: transformStyleObjectToString(node.style),
+          }, children)
+        },
+      },
+    }
+    expect(slateToHtml(slate, config)).toEqual(html)
+  })
+
+  it('mark transforms', () => {
+    const html = '<p><strong style="font-size: 96px; --text-color: #DD3A0A; @media screen { z-index: 1; color: var(--text-color) }">Paragraph</strong></p>'
+    const slate = [
+      {
+        type: 'p',
+        children: [
+          {
+            bold: true,
+            style: {
+              fontSize: '96px',
+              '--text-color': '#DD3A0A',
+              '@media screen': {
+                zIndex: '1',
+                color: 'var(--text-color)'
+              }
+            },
+            text: 'Paragraph',
+          },
+        ],
+      }
+    ]
+    const config = {
+      ...slateToDomConfig,
+      markTransforms: {
+        ...slateToDomConfig.markTransforms,
+        strong: ({ node }: { node?: any }) => {
+          return new Element('strong', {
+            style: transformStyleObjectToString(node.style),
+          })
+        },
+      },
+    }
+    expect(slateToHtml(slate, config)).toEqual(html)
+  })
+
+  it('mark transforms on multiple marks', () => {
+    const html =
+      '<p>This is editable <strong style=\"font-size: 20px; font-weight: 600; text-decoration: underline dotted\">rich</strong> text, <i style=\"text-decoration: underline\">much</i> better than a <pre><code style=\"color: red\">&lt;textarea&gt;</code></pre>!</p>'
+    const slate = [
+      {
+        type: 'p',
+        children: [
+          {
+            text: 'This is editable ',
+          },
+          {
+            text: 'rich',
+            bold: true,
+            style: {
+              fontSize: "20px",
+              fontWeight: 600,
+              textDecoration: "underline dotted",
+            }
+          },
+          {
+            text: ' text, ',
+          },
+          {
+            text: 'much',
+            italic: true,
+            style: {
+              textDecoration: "underline",
+            }
+          },
+          {
+            text: ' better than a ',
+          },
+          {
+            text: '<textarea>',
+            code: true,
+            style: {
+              color: "red",
+            }
+          },
+          {
+            text: '!',
+          },
+        ],
+      }
+    ]
+    const config = {
+      ...slateToDomConfig,
+      markTransforms: {
+        ...slateToDomConfig.markTransforms,
+        code: ({ node }: { node?: any }) => {
+          return new Element('code', {
+            style: transformStyleObjectToString(node.style),
+          })
+        },
+        i: ({ node }: { node?: any }) => {
+          return new Element('i', {
+            style: transformStyleObjectToString(node.style),
+          })
+        },
+        strong: ({ node }: { node?: any }) => {
+          return new Element('strong', {
+            style: transformStyleObjectToString(node.style),
           })
         },
       },