+++ __tests__/utilities/domhandler.spec.ts
@@ -1,3 +1,4 @@
+import { Element } from "domhandler"
 import { nestedMarkElementsString } from '../../src/utilities/domhandler'
 
 /**
@@ -6,13 +7,13 @@ import { nestedMarkElementsString } from '../../src/utilities/domhandler'
 
 describe('nestedMarkElementsString', () => {
   it('nests text in 1 mark element', () => {
-    expect(nestedMarkElementsString(['u'], 'Unarticulated Annotation text')).toEqual(
+    expect(nestedMarkElementsString([new Element('u', {})], 'Unarticulated Annotation text')).toEqual(
       '<u>Unarticulated Annotation text</u>',
     )
   })
 
   it('nests text in 2 mark elements', () => {
-    expect(nestedMarkElementsString(['u', 'i'], 'Unarticulated Annotation, and Idiomatic Text text')).toEqual(
+    expect(nestedMarkElementsString([new Element('u', {}), new Element('i', {})], 'Unarticulated Annotation, and Idiomatic Text text')).toEqual(
       '<u><i>Unarticulated Annotation, and Idiomatic Text text</i></u>',
     )
   })
@@ -20,7 +21,7 @@ describe('nestedMarkElementsString', () => {
   it('nests text in 3 mark elements', () => {
     expect(
       nestedMarkElementsString(
-        ['u', 'i', 'strong'],
+        [new Element('u', {}), new Element('i', {}), new Element('strong', {})],
         'Strong Importance, Unarticulated Annotation, and Idiomatic Text text',
       ),
     ).toEqual('<u><i><strong>Strong Importance, Unarticulated Annotation, and Idiomatic Text text</strong></i></u>')
@@ -29,7 +30,7 @@ describe('nestedMarkElementsString', () => {
   it('nests text in 4 mark elements', () => {
     expect(
       nestedMarkElementsString(
-        ['u', 'i', 'strong', 's'],
+        [new Element('u', {}), new Element('i', {}), new Element('strong', {}), new Element('s', {})],
         'Strong Importance, Unarticulated Annotation, Idiomatic Text and Strikethrough text',
       ),
     ).toEqual(