+++ src/config/slateToDom/types.ts
@@ -1,5 +1,15 @@
 import { ChildNode, Element } from 'domhandler'
 
+interface MarkTagTransform {
+  [key: string]: ({
+    node,
+    attribs,
+  }: {
+    node?: any
+    attribs?: { [key: string]: string }
+  }) => Element
+}
+
 interface ElementTagTransform {
   [key: string]: ({
     node,
@@ -18,9 +28,14 @@ export interface Config {
   elementStyleMap?: {
     [key: string]: string
   }
+  markStyleMap?: {
+    [key: string]: string
+  }
+  markTransforms?: MarkTagTransform
   elementTransforms: ElementTagTransform
   defaultTag?: string
   encodeEntities?: boolean
+  alwaysEncodeBreakingEntities?: boolean
   alwaysEncodeCodeEntities?: boolean
   convertLineBreakToBr?: boolean
 }